pyproject.toml
[tool]
[tool.poetry]
name = "mongoengine_dsl"
version = "0.1.0"
homepage = "https://github.com/StoneMoe/mongoengine_dsl"
description = "DSL to MongoEngine Q."
authors = ["Lake Chan <lakechan96@gmail.com>"]
readme = "README.md"
license = "MIT"
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "mongoengine_dsl" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.6.1,<4.0"

# deps
mongoengine = { version = "~0" }
antlr4-python3-runtime = { version = "4.9.2" }
# extra - test
pytest = { version = "6.1.2", optional = true }
pytest-cov = { version = "2.10.1", optional = true }
black = { version = "20.8b1", optional = true }
isort = { version = "5.6.4", optional = true }
flake8 = { version = "3.8.4", optional = true }
flake8-docstrings = { version = "^1.6.0", optional = true }
mongomock = { version = "3.23.0", optional = true }
# extra - dev
tox = { version = "^3.20.1", optional = true }
pre-commit = { version = "^2.12.0", optional = true }
virtualenv = { version = "^20.2.2", optional = true }
pip = { version = "^20.3.1", optional = true }
twine = { version = "^3.3.0", optional = true }
toml = { version = "^0.10.2", optional = true }
# extra - doc
mkdocs = { version = "^1.1.2", optional = true }
mkdocs-include-markdown-plugin = { version = "^1.0.0", optional = true }
mkdocs-material = { version = "^6.1.7", optional = true }
mkdocstrings = { version = "^0.13.6", optional = true }
mkdocs-material-extensions = { version = "^1.0.1", optional = true }
mkdocs-autorefs = { version = "0.1.1", optional = true }
livereload = { version = "2.6.3", optional = true }

[tool.poetry.extras]
test = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "flake8-docstrings",
    "mongomock"
]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs",
    "livereload"
]



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
